// src/data/openapi/types.gen.ts
export type TeamDto = {
    id?: number;
    name?: string;
    city?: string;
    logoColor?: string;
    players?: Array<TeamPlayerDto>;
};

export type TeamPlayerDto = {
    id?: number;
    name?: string;
    roles?: Array<'R4' | 'Central' | 'Passeur' | 'Pointu' | 'Libero'>;
    numero?: number;
    age?: number;
    taille?: string;
};